refactor: move search menu to component
