+++ src/components/SearchMenu/constants.js
@@ -1,0 +1,6 @@
+export const PHASE_TYPE = {
+  AI: 'ai',
+  NORMAL: 'normal'
+}
+
+export default {}
